requirements.txt
streamlit
geopandas
rasterio
numpy
earthengine-api
google-auth
